健康情報CSV, Excelはmodelとかをbusiness配下に移動
