+++ ha-business/src/main/java/jp/co/ha/business/io/file/csv/reader/HealthInfoCsvReader.java
@@ -1,7 +1,7 @@
-package jp.co.ha.web.file.csv.reader;
+package jp.co.ha.business.io.file.csv.reader;
 
+import jp.co.ha.business.io.file.csv.model.HealthInfoCsvUploadModel;
 import jp.co.ha.common.io.file.csv.reader.CsvReader;
-import jp.co.ha.web.file.csv.model.HealthInfoCsvUploadModel;
 
 public class HealthInfoCsvReader extends CsvReader<HealthInfoCsvUploadModel> {
