+++ ha-business/src/main/java/jp/co/ha/business/io/file/excel/model/HealthInfoExcelModel.java
@@ -1,18 +1,16 @@
-package jp.co.ha.web.file.excel.model;
-
-import java.util.Date;
+package jp.co.ha.business.io.file.excel.model;
 
 import jp.co.ha.common.io.file.excel.annotation.ExcelDownloadModel;
 import jp.co.ha.common.io.file.excel.model.BaseExcelModel;
 
 /**
- * 結果照会Excel出力モデルクラス<br>
+ * 健康情報Excel出力モデルクラス<br>
  *
  */
-@ExcelDownloadModel(sheetName = "健康情報照会"
-					, headerNames = { "身長", "体重", "BMI", "標準体重", "登録日時" }
-					, footerNames = { "身長", "体重", "BMI", "標準体重", "登録日時" })
-public class ReferenceExcelModel implements BaseExcelModel {
+@ExcelDownloadModel(sheetName = "健康情報"
+					, headerNames = { "身長", "体重", "BMI", "標準体重" }
+					, footerNames = { "身長", "体重", "BMI", "標準体重" } )
+public class HealthInfoExcelModel implements BaseExcelModel {
 
 	/** 身長 */
 	private String height;
@@ -22,8 +20,6 @@ public class ReferenceExcelModel implements BaseExcelModel {
 	private String bmi;
 	/** 標準体重 */
 	private String standardWeight;
-	/** 登録日時 */
-	private Date regDate;
 
 	/**
 	 * heightを返す
@@ -99,25 +95,6 @@ public class ReferenceExcelModel implements BaseExcelModel {
 	 */
 	public void setStandardWeight(String standardWeight) {
 		this.standardWeight = standardWeight;
-	}
-
-	/**
-	 * regDateを返す<br>
-	 *
-	 * @return regDate regDate
-	 */
-	public Date getRegDate() {
-		return regDate;
-	}
-
-	/**
-	 * regDateを設定する<br>
-	 *
-	 * @param regDate
-	 *     regDate
-	 */
-	public void setRegDate(Date regDate) {
-		this.regDate = regDate;
 	}
 
 }
