+++ ha-web/src/main/java/jp/co/ha/web/service/HealthInfoReferenceService.java
@@ -3,8 +3,8 @@ package jp.co.ha.web.service;
 import java.util.List;
 
 import jp.co.ha.api.response.HealthInfoReferenceResponse;
+import jp.co.ha.business.io.file.csv.model.ReferenceCsvDownloadModel;
 import jp.co.ha.common.exception.BaseException;
-import jp.co.ha.web.file.csv.model.ReferenceCsvDownloadModel;
 import jp.co.ha.web.form.HealthInfoReferenceForm;
 
 /**
