+++ ha-web/src/main/java/jp/co/ha/web/service/impl/HealthInfoCsvUploadServiceImpl.java
@@ -8,13 +8,13 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import jp.co.ha.business.io.file.csv.model.HealthInfoCsvUploadModel;
+import jp.co.ha.business.io.file.csv.reader.HealthInfoCsvReader;
 import jp.co.ha.common.exception.AppIOException;
 import jp.co.ha.common.exception.BaseException;
 import jp.co.ha.common.exception.ErrorCode;
 import jp.co.ha.common.io.file.csv.reader.CsvReader;
 import jp.co.ha.common.io.file.csv.service.CsvUploadService;
-import jp.co.ha.web.file.csv.model.HealthInfoCsvUploadModel;
-import jp.co.ha.web.file.csv.reader.HealthInfoCsvReader;
 
 /**
  * 健康情報ファイルアップロードサービス実装クラス<br>
